[Feat/#36]-review 삭제 이전 코드 삭제 및 set 코드로 다시 구성
